// src/company.d.ts
export interface CompanySearch {
    currency: string
    exchangeShortName: string
    name: string
    stockExchange: string
    symbol: string
}

export interface CompanyProfile {
    symbol: string
    price: number
    beta: number
    marketCap: number
    lastDividend: number
    range: string
    change: number
    changePercentage: number
    volume: number
    averageVolume: number
    companyName: string
    currency: string
    cik: string
    isin: string
    cusip: string
    exchangeFullName: string
    exchange: string
    industry: string
    website: string
    description: string
    ceo: string
    sector: string
    country: string
    fullTimeEmployees: string
    phone: string
    address: string
    city: string
    state: string
    zip: string
    image: string
    ipoDate: string
    defaultImage: boolean
    isEtf: boolean
    isActivelyTrading: boolean
    isAdr: boolean
    isFund: boolean
    // Legacy/v3 fields retained as optional for compatibility.
    volAvg?: number
    mktCap?: number
    lastDiv?: number
    changes?: number
    exchangeShortName?: string
    counter?: string
    dcfDiff?: number
    dcf?: number
}
// ...
export interface CompanyIncomeStatement {
    date: string
    symbol: string
    reportedCurrency: string
    cik: string
    fillingDate: string
    acceptedDate: string
    calendarYear: string
    period: string
    revenue: number
    costOfRevenue: number
    grossProfit: number
    grossProfitRatio: number
    researchAndDevelopmentExpenses: number
    generalAndAdministrativeExpenses: number
    sellingAndMarketingExpenses: number
    sellingGeneralAndAdministrativeExpenses: number
    otherExpenses: number
    operatingExpenses: number
    costAndExpenses: number
    interestIncome: number
    interestExpense: number
    depreciationAndAmortization: number
    ebitda: number
    ebitdaratio: number
    operatingIncome: number
    operatingIncomeRatio: number
    totalOtherIncomeExpensesNet: number
    incomeBeforeTax: number
    incomeBeforeTaxRatio: number
    incomeTaxExpense: number
    netIncome: number
    netIncomeRatio: number
    eps: number
    epsdiluted: number
    weightedAverageShsOut: number
    weightedAverageShsOutDil: number
    link: string
    finalLink: string
}

export interface CompanyBalanceSheet {
    date: string
    symbol: string
    reportedCurrency: string
    cik: string
    fillingDate: string
    acceptedDate: string
    calendarYear: string
    period: string
    cashAndCashEquivalents: number
    shortTermInvestments: number
    cashAndShortTermInvestments: number
    netReceivables: number
    inventory: number
    otherCurrentAssets: number
    totalCurrentAssets: number
    propertyPlantEquipmentNet: number
    goodwill: number
    intangibleAssets: number
    goodwillAndIntangibleAssets: number
    longTermInvestments: number
    taxAssets: number
    otherNonCurrentAssets: number
    totalNonCurrentAssets: number
    otherAssets: number
    totalAssets: number
    accountPayables: number
    shortTermDebt: number
    taxPayables: number
    deferredRevenue: number
    otherCurrentLiabilities: number
    totalCurrentLiabilities: number
    longTermDebt: number
    deferredRevenueNonCurrent: number
    deferredTaxLiabilitiesNonCurrent: number
    otherNonCurrentLiabilities: number
    totalNonCurrentLiabilities: number
    otherLiabilities: number
    capitalLeaseObligations: number
    totalLiabilities: number
    preferredStock: number
    commonStock: number
    retainedEarnings: number
    accumulatedOtherComprehensiveIncomeLoss: number
    othertotalStockholdersEquity: number
    totalStockholdersEquity: number
    totalEquity: number
    totalLiabilitiesAndStockholdersEquity: number
    minorityInterest: number
    totalLiabilitiesAndTotalEquity: number
    totalInvestments: number
    totalDebt: number
    netDebt: number
    link: string
    finalLink: string
}

export interface CompanyCashflow {
    date: string
    symbol: string
    reportedCurrency: string
    cik: string
    fillingDate: string
    acceptedDate: string
    calendarYear: string
    period: string
    netIncome: number
    depreciationAndAmortization: number
    deferredIncomeTax: number
    stockBasedCompensation: number
    changeInWorkingCapital: number
    accountsReceivables: number
    inventory: number
    accountsPayables: number
    otherWorkingCapital: number
    otherNonCashItems: number
    netCashProvidedByOperatingActivities: number
    investmentsInPropertyPlantAndEquipment: number
    acquisitionsNet: number
    purchasesOfInvestments: number
    salesMaturitiesOfInvestments: number
    otherInvestingActivites: number
    netCashUsedForInvestingActivites: number
    debtRepayment: number
    commonStockIssued: number
    commonStockRepurchased: number
    dividendsPaid: number
    otherFinancingActivites: number
    netCashUsedProvidedByFinancingActivities: number
    effectOfForexChangesOnCash: number
    netChangeInCash: number
    cashAtEndOfPeriod: number
    cashAtBeginningOfPeriod: number
    operatingCashFlow: number
    capitalExpenditure: number
    freeCashFlow: number
    link: string
    finalLink: string
}

export interface CompanyKeyMetrics {
    revenuePerShareTTM: number
    netIncomePerShareTTM: number
    operatingCashFlowPerShareTTM: number
    freeCashFlowPerShareTTM: number
    cashPerShareTTM: number
    bookValuePerShareTTM: number
    tangibleBookValuePerShareTTM: number
    shareholdersEquityPerShareTTM: number
    interestDebtPerShareTTM: number
    marketCapTTM: number
    enterpriseValueTTM: number
    peRatioTTM: number
    priceToSalesRatioTTM: number
    pocfratioTTM: number
    pfcfRatioTTM: number
    pbRatioTTM: number
    ptbRatioTTM: number
    evToSalesTTM: number
    enterpriseValueOverEBITDATTM: number
    evToOperatingCashFlowTTM: number
    evToFreeCashFlowTTM: number
    earningsYieldTTM: number
    freeCashFlowYieldTTM: number
    debtToEquityTTM: number
    debtToAssetsTTM: number
    netDebtToEBITDATTM: number
    currentRatioTTM: number
    interestCoverageTTM: number
    incomeQualityTTM: number
    dividendYieldTTM: number
    dividendYieldPercentageTTM: number
    payoutRatioTTM: number
    salesGeneralAndAdministrativeToRevenueTTM: number
    researchAndDevelopementToRevenueTTM: number
    intangiblesToTotalAssetsTTM: number
    capexToOperatingCashFlowTTM: number
    capexToRevenueTTM: number
    capexToDepreciationTTM: number
    stockBasedCompensationToRevenueTTM: number
    grahamNumberTTM: number
    roicTTM: number
    returnOnTangibleAssetsTTM: number
    grahamNetNetTTM: number
    workingCapitalTTM: number
    tangibleAssetValueTTM: number
    netCurrentAssetValueTTM: number
    investedCapitalTTM: number
    averageReceivablesTTM: number
    averagePayablesTTM: number
    averageInventoryTTM: number
    daysSalesOutstandingTTM: number
    daysPayablesOutstandingTTM: number
    daysOfInventoryOnHandTTM: number
    receivablesTurnoverTTM: number
    payablesTurnoverTTM: number
    inventoryTurnoverTTM: number
    roeTTM: number
    capexPerShareTTM: number
    dividendPerShareTTM: number
    debtToMarketCapTTM: number
}

export interface CompanyHistoricalDividend {
    symbol: string
    historical: Dividend[]
}

export interface Dividend {
    date: string
    label: string
    adjDividend: number
    dividend: number
    recordDate: string
    paymentDate: string
    declarationDate: string
}

export interface CompanyComparisonData {
    symbol: string
    peersList: string[]
}

export interface CompanyTenK {
    symbol: string
    fillingDate: string
    acceptedDate: string
    cik: string
    type: string
    link: string
    finalLink: string
}